Replace UI Kit Button
