+++ apps/onboarding/src/components/DebuggerPage/DebuggerPage.types.ts
@@ -7,12 +7,11 @@ type InputBase = {
 
 type InputText = InputBase & { type: 'text' }
 type InputNumber = InputBase & { type: 'number' }
-type InputStepper = InputBase & { type: 'stepper' }
 type InputSelect = InputBase & { type: 'select'; options: Array<{ name: string; value: string }> }
 type InputEmail = InputBase & { type: 'email' }
 type InputDate = InputBase & { type: 'date' }
 
-export type Input = InputText | InputNumber | InputStepper | InputSelect | InputEmail | InputDate
+export type Input = InputText | InputNumber | InputSelect | InputEmail | InputDate
 
 export type InputGroup = {
   title: string
