+++ apps/onboarding/src/components/ForeverPage/ForeverCodeReadyPage.tsx
@@ -1,7 +1,8 @@
 import { keyframes } from '@emotion/react'
 import styled from '@emotion/styled'
 import { useTranslation } from 'next-i18next'
-import { LinkButton, Space } from 'ui'
+import { Space } from 'ui'
+import { LinkButton } from '@/components/Button/Button'
 import { useCurrentLocale } from '@/lib/l10n'
 import { PageLink } from '@/lib/PageLink'
 import { PageLayout } from './PageLayout'
