+++ apps/onboarding/src/components/ForeverPage/ForeverPage.tsx
@@ -1,7 +1,8 @@
 import styled from '@emotion/styled'
 import { useTranslation } from 'next-i18next'
 import { useRouter } from 'next/router'
-import { Button, InputField, Space } from 'ui'
+import { InputField, Space } from 'ui'
+import { Button } from '@/components/Button/Button'
 import { useForm } from '@/hooks/useForm'
 import { PageLayout } from './PageLayout'
 import { usePrintCodeEffect } from './usePrintCodeEffect'
