+++ apps/onboarding/src/components/LandingPage/LandingPage.tsx
@@ -2,7 +2,8 @@ import styled from '@emotion/styled'
 import { useTranslation } from 'next-i18next'
 import { useRouter } from 'next/router'
 import { ChangeEvent, useCallback, useState, useMemo } from 'react'
-import { Button, mq, theme } from 'ui'
+import { mq, theme } from 'ui'
+import { Button } from '@/components/Button/Button'
 import { Header } from '@/components/Nav/Header'
 import { useCurrentLocale } from '@/lib/l10n'
 import { PageLink } from '@/lib/PageLink'
