+++ apps/onboarding/src/components/StartPage/StartPage.tsx
@@ -3,7 +3,8 @@ import styled from '@emotion/styled'
 import * as RadioGroup from '@radix-ui/react-radio-group'
 import { useTranslation } from 'next-i18next'
 import { useState } from 'react'
-import { Button, Space } from 'ui'
+import { Space } from 'ui'
+import { Button } from '@/components/Button/Button'
 import { HeroImage } from '@/components/hero-image'
 import { PageHeaderLayout } from '@/components/page-header-layout'
 import { useForm } from '@/hooks/useForm'
