+++ apps/store/src/components/ProductPage/PurchaseForm/PurchaseForm.tsx
@@ -118,9 +118,7 @@ const PendingState = () => {
   return (
     <SectionWrapper>
       <OpenModalButtonWrapper>
-        <Button disabled fullWidth>
-          {t('OPEN_PRICE_CALCULATOR_BUTTON')}
-        </Button>
+        <Button disabled>{t('OPEN_PRICE_CALCULATOR_BUTTON')}</Button>
       </OpenModalButtonWrapper>
     </SectionWrapper>
   )
@@ -132,11 +130,7 @@ const IdleState = ({ onClick }: IdleStateProps) => {
   const ref = useRef<HTMLDivElement>(null)
   const { t } = useTranslation('purchase-form')
 
-  const button = (
-    <Button onClick={onClick} fullWidth>
-      {t('OPEN_PRICE_CALCULATOR_BUTTON')}
-    </Button>
-  )
+  const button = <Button onClick={onClick}>{t('OPEN_PRICE_CALCULATOR_BUTTON')}</Button>
 
   return (
     <>
