+++ packages/ui/src/index.ts
@@ -1,17 +1,13 @@
 export { Checkbox } from './components/Checkbox/Checkbox'
 export type { CheckboxProps } from './components/Checkbox/Checkbox'
-export { BurgerButton } from './components/BurgerButton/BurgerButton'
 export { HedvigLogo } from './components/HedvigLogo/HedvigLogo'
-export type { ButtonProps } from './components/Button/Button'
-export { Button, LinkButton } from './components/Button/Button'
-export type { ButtonVariant, ButtonColors, ButtonSize } from './components/Button/Button'
 export { Separate } from './components/Separate'
 export { Space } from './components/Space'
+export { Button } from './components/Button/Button'
 export { InputBase } from './components/InputBase'
 export type { InputBaseProps } from './components/InputBase'
 export { InputField } from './components/InputField/InputField'
 export type { InputFieldProps } from './components/InputField/InputField'
-export { InputStepper } from './components/InputStepper/InputStepper'
 export { Heading } from './components/Heading/Heading'
 export type { HeadingProps } from './components/Heading/Heading'
 export { Text } from './components/Text/Text'
